adjust trait bound for with_clone::Any
I believe the [new] std::marker::[Reflect] trait made it necessary to
specify a Reflect bound, but it's advised to use an Any bound instead.

Adding this bound makes it compile. I ran the tests and got an error
saying it couldn't determine the type of one of the expressions so I
introduced a scope to bind the result of the expression and annotate its
type.

I also removed the `convert` feature that was no longer needed.

[new]: https://github.com/rust-lang/rust/pull/23712
[Reflect]: http://doc.rust-lang.org/std/marker/trait.Reflect.html